feat(player): Add getPlayer

+++ src/main/kotlin/com/cnicolas/rpg/dungeondragonsserver/database/entities/Player.kt
@@ -6,5 +6,5 @@ import javax.persistence.*
 @Table
 data class Player(
     var name: String = "",
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val id: Int
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY) val id: Long
 )
